+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+- **Fixed:** Placeholder text from schema. [\#152](https://github.com/vazco/uniforms/issues/152)
+
+<br>
+
 ## [v1.8.0](https://github.com/vazco/uniforms/releases/tag/v1.8.0)
 - **Added:** Support for direct imports, i.e. `import connectField from 'uniforms'`. [\#129](https://github.com/vazco/uniforms/issues/129)
 - **Fixed:** Optional packages resolution in Webpack. [\#150](https://github.com/vazco/uniforms/issues/150)